Add support for RFC6750 Bearer Tokens.
Bearer tokens as defined in RFC6570 work very similarly to the way tokens are defined in grails-spring-security-rest.  With this update, a configuration setting can be turned on which will make the filter look for authentication "tokens" to be transmitted in the methods defined by the specification.  The filter will expect to see the token in the Authorization header as "Bearer <token>", or in the "access_token" query string parameter.

Because of design considerations, the third method of transmitting bearer tokens via form request body has not been implemented.